Split up project into separate class files that can compile
